phonebook/requirements.txt
grpcio
grpcio-tools
mypy
dotenv
colorama